+++ frontend/src/App.tsx
@@ -1,12 +1,12 @@
 import { Routes, Route } from "react-router-dom";
 import Home from "@/pages/Home";
-import Login from "@/pages/Login"; // create this page
+import SighUp from "@/pages/SignUp"; // create this page
 
 function App() {
   return (
     <Routes>
       <Route path="/" element={<Home />} />
-      <Route path="/login" element={<Login />} />
+      <Route path="/signup" element={<SighUp />} />
     </Routes>
   );
 }
