+++ frontend/src/pages/SignUp.tsx
@@ -1,9 +1,9 @@
-import LoginForm from "@/components/LoginForm"; // your Spotify login form
+import SignIn from "@/components/SpotifyAuthForm"; 
 
 const Login = () => {
   return (
     <div className="min-h-screen flex items-center justify-center bg-gray-50 p-6">
-      <LoginForm />
+      <SignIn />
     </div>
   );
 };
